Revert "added shacn and improved home page"
