+++ package.json
@@ -10,7 +10,6 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@radix-ui/react-icons": "^1.3.0",
     "@fortawesome/fontawesome-svg-core": "^6.6.0",
     "@fortawesome/free-brands-svg-icons": "^6.6.0",
     "@fortawesome/free-regular-svg-icons": "^6.6.0",
@@ -19,31 +18,25 @@
     "@reduxjs/toolkit": "^2.0.1",
     "@tinymce/tinymce-react": "^4.3.2",
     "appwrite": "^13.0.1",
-    "class-variance-authority": "^0.7.0",
-    "clsx": "^2.1.1",
     "html-react-parser": "^5.1.1",
-    "lucide-react": "^0.447.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.49.3",
     "react-redux": "^9.1.0",
     "react-router-dom": "^6.21.2",
-    "tailwind-merge": "^2.5.2",
-    "tailwindcss-animate": "^1.0.7"
     "react-share": "^5.1.0"
   },
   "devDependencies": {
-    "@types/node": "^22.7.4",
     "@types/react": "^18.2.43",
     "@types/react-dom": "^18.2.17",
     "@vitejs/plugin-react": "^4.2.1",
-    "autoprefixer": "^10.4.20",
+    "autoprefixer": "^10.4.16",
     "eslint": "^8.55.0",
     "eslint-plugin-react": "^7.33.2",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.5",
-    "postcss": "^8.4.47",
-    "tailwindcss": "^3.4.13",
+    "postcss": "^8.4.33",
+    "tailwindcss": "^3.4.1",
     "vite": "^5.0.8"
   }
 }
