+++ src/App.jsx
@@ -2,8 +2,8 @@ import React, { useState, useEffect } from 'react'
 import { useDispatch } from 'react-redux'
 import './App.css'
 import authService from "./appwrite/auth"
-import { login, logout } from "./store/authSlice"
-import { Footer, Header } from './components'
+import {login, logout} from "./store/authSlice"
+import { Footer, Header } from './componenets'
 import { Outlet } from 'react-router-dom'
 
 function App() {
@@ -12,24 +12,22 @@ function App() {
 
   useEffect(() => {
     authService.getCurrentUser()
-      .then((userData) => {
-        if (userData) {
-          dispatch(login({ userData }))
-        } else {
-          dispatch(logout())
-        }
-      })
-      .finally(() => setLoading(false))
+    .then((userData) => {
+      if (userData) {
+        dispatch(login({userData}))
+      } else {
+        dispatch(logout())
+      }
+    })
+    .finally(() => setLoading(false))
   }, [])
-
+  
   return !loading ? (
-    <div className='bg-gray-100'>
-      <div className=''>
+    <div className='min-h-screen flex flex-wrap content-between bg-gray-400'>
+      <div className='w-full block'>
         <Header />
         <main>
-          <div className='text-center'>
-            TODO:  <Outlet />
-          </div>
+        TODO:  <Outlet />
         </main>
         <Footer />
       </div>
