+++ src/componenets/Header/Header.jsx
@@ -38,7 +38,7 @@ function Header() {
 
 
   return (
-    <header className='py-2 shadow bg-gray-200'>
+    <header className='py-3 shadow bg-gray-500'>
       <Container>
         <nav className='flex'>
           <div className='mr-4'>
