+++ src/componenets/Logo.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function Logo({width = '100px'}) {
+  return (
+    <div>
+     Logo
+    </div>
+  )
+}
+
+export default Logo
