+++ src/pages/Login.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { Login as LoginComponent } from '../components'
+import { Login as LoginComponent } from '../componenets'
 
 function Login() {
   return (
