+++ src/pages/Post.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import { Link, useNavigate, useParams } from "react-router-dom";
 import service from "../appwrite/configAppwrite";
-import { Button, Container } from "../components";
+import { Button, Container } from "../componenets";
 import parse from "html-react-parser";
 import { useSelector } from "react-redux";
 import {
